Improve printout when Jenkins master is not configured
If a Jenkins where the URL has not been set responds to the UDP
broadcast the following will be printed on the command line:

/10.48.20.70 doesn't have the configuration set yet. Please go to the system configuration page of this Jenkins and submit it: <hudson><version>1.596.2</version><server-id>83183a365760ffeda7fc764f33745c96</server-id><slave-port>59561</slave-port><swarm>9838a12a-4448-4b0d-a810-e6161791b050</swarm></hudson>

Instead print what actually needs to be set. Since we don't know what
the port of the Jenkins server is we can't generate the full
URL. Example print out:

Jenkins master at '/10.48.20.70' doesn't have a valid Jenkins URL configuration set. Please go to <jenkins url>/configure and set a valid URL.