Improve end user reporting of hostname lookup errors
The CanonicalHostName() method call is only best effors as stated in the
Java documentation:

   "Gets the fully qualified domain name for this IP address. Best effort
   method, meaning we may not be able to return the FQDN depending on
   the underlying system configuration"

This means that we can experience failures if the system configuration
is such that when the swarm client is launched we can't determine our
hostname. Since swarm slaves are often spawned in public/private clouds
there might be some eventual consistency with regards to lookup of
hostnames. We try to help the situation by printing a somewhat
informative error message.

Using the '-name' command line option for this should have been a way to
work around this. However we where unconditionally performing this
lookup. Instead we now only perform the lookup of no name has been
provided on the command line.

Thanks to @jacob-keller who reported this.