Correct Dockerfiles to actually work
Make sure when we run docker-compose up that we actually get the
containers in a state that they are ready to be interconnected.

Right now one needs to run the bash script to upload the newly built
swarm plugin.

TODO: Add a simple jenkins job that executes on the swarm connected
agent so we are 100% sure it will work properly.