Add option for -pidFile
When running the swarm client as a daemon or with a tools like
systemd, it is useful to have a file with the process ID. This commit
introduces a new option in the client which will generate that file
when the client starts up. The file will be deleted when the process
exits.