Rename remoteFsRoot variable -> fsroot (#337)
Previously, this was the only option which had an internal variable
name which differed from the exposed option name. This could be
confusing for users wanting to use the new YAML config file format,
since it meant that they would need to specify `remoteFsRoot` there,
and look at the source code to figure out this option name.