Check path of old PID (if running) before exiting
The commit a982063 changed the behavior of the `-pidFile` argument so
that the client would exit if a running process was found with the PID
listed in the PID file. This was meant to prevent multiple instances of
Swarm Client from being started (namely when Swarm Client is launched
without a proper service manager that would prevent multiple instances
from running).

However, this behavior introduced a dangerous flaw: if Swarm Client does
not exit cleanly (because of a sudden reboot, crash, etc) and another
process starts with this PID, then Swarm Client will not start at all.

This commit adds a small safety check by comparing the paths of the old
and current process. If they are the same, we can assume that the old
process is another Swarm Client instance and bail out. But if the paths
don't match, we should instead consider the PID file to be stale and
continue startup as usual.